Add files via upload
this branch focuses on saving using a reset and replay function. this eases the strain of having to figure out how to get a save function that will delete picked up items that are initialized on the map when the game is loaded fresh. a couple concerns with this method is save file size and load time as the game is essentially replaying the game very quickly. will be optimized as much as possible